requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
pyotp==2.9.0
cryptography==41.0.7
slowapi==0.1.9
email-validator==2.2.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
alembic==1.16.1
gunicorn==23.0.0
openai==1.99.9
